adding kubernetes and docker

+++ scripts/docker.sh
@@ -1,12 +1,14 @@
 #!/usr/bin/env bash
 
-echo "Install docker"
+# Installing docker if bin does not exist
+if [ ! -f /usr/bin/docker ]; then
+  echo "Installing docker"
+  yum install -y yum-utils device-mapper-persistent-data lvm2
+  yum-config-manager --add-repo https://download.docker.com/linux/centos/docker-ce.repo
+  yum -y install docker-ce-18.06.1.ce-3.el7
 
-yum install -y yum-utils device-mapper-persistent-data lvm2
-yum-config-manager --add-repo https://download.docker.com/linux/centos/docker-ce.repo
-yum -y install docker-ce
+  usermod -aG docker vagrant
 
-usermod -aG docker vagrant
-
-systemctl enable docker
-systemctl restart docker
+  systemctl enable docker
+  systemctl restart docker
+fi
